Add support for filtering sensitive information before it reaches the logs.
When rdr_XXX_sensitive() functions are used for logging, it is now
possible to filter certain texts in the messages depending on a global
flag (default off - no filtering).

When filtering is enabled the texts put between curly brackets would be
replaced with # symbol. For example

rdr_log_sensitive(reader, "serial: {%s}", "123 456 789");

would result in the following entries:

... serial: 123 456 789 (when filtering is disabled)
... serial: ########### (when filtering is enabled)

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@7129 9dab235a-647b-404a-b20e-d39061ee4083