cfg: Change int variable types to the ones before config parser.
The variable types that were {u}int8_t were changed to {u}int32_t
when conversion to table based config parser was made because
the parser lacked support for smaller types.

Now that the support is added, lets return the old types and stop
wasting memory. A lot of the variables really should be changed to
"bool" but that can be made later.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@7514 9dab235a-647b-404a-b20e-d39061ee4083