Fix all warnings about shadowed variables reported by clang.
Shadowed variables can lead to very subtle and hard to find bugs
so it is a good idea to fix them.

Unfortunately -Wshadow can not be enabled in the Makefile because
gcc is brain damaged and warns about variables with names like
built-in functions. So if anybody wants to test his code, better
use clang for this kind of checking otherwise you'll be swamped with
stupid useless warnings.

The magic command is: `make CC=clang EXTRA_CC_WARN=-Wshadow`

This commit sort of fixes ticket #2823.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@7611 9dab235a-647b-404a-b20e-d39061ee4083