Enable -Wshadow warning messages.
Enable -Wshadow warning messages and fix all variables that are
named the same way as built-in variables. That reduces the noise
when grepping for certain function. -Wshadow is generally a good
option that may prevent subtle code errors.

This is the full fix for ticket #2823.

There should be no behaviour changes after this commit.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@7614 9dab235a-647b-404a-b20e-d39061ee4083