Replace fast_rnd() with get_random_bytes().
All of the places that used fast_rnd() really wanted more than one
random byte so get_random_bytes() is better fit for the usage. Additional
extra is that the way get_random_bytes() is implemented it is relatively
secure on platforms that provide either /dev/urandom or /dev/random.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@7655 9dab235a-647b-404a-b20e-d39061ee4083