Revert "According to the GNU Make manual, the flags are automatically parsed to"
This reverts commit r7970.

The purpose of `$(MAKEFLAGS)` in Makefile.extra is to allow overwriting
variables that are already set on the make command line. See the difference:

Without `$(MAKEFLAGS)` (commit r7970):

{{{
$ make static EXTRA_TARGET=""
make --no-print-directory \
        EXTRA_FLAGS=-static \
        EXTRA_TARGET=-static
}}}

With `$(MAKEFLAGS)` (after reverting r7970):

{{{
$ make static EXTRA_TARGET=""
make --no-print-directory \
        EXTRA_FLAGS=-static \
        EXTRA_TARGET=-static \
        EXTRA_TARGET=
}}}

Yes, `$(MAKEFLAGS)` are automatically passed to sub-makes but they
are put before any variables that are set on the command line. This prevents
reseting already set variables.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@7971 9dab235a-647b-404a-b20e-d39061ee4083