build: Prevent LD from leaving temp files in /tmp.
The LD output is saved into variable and then processed, because if
the output is piped directly into another command LD creates 4 files
in your /tmp directory and doesn't delete them.

Also remove linker version in info output, its useless and again
there was a problem with the temp files.

Because of this LD behaviour each running of `make` left you with
at least 8 files in your /tmp directory which was highly annoying.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@8329 9dab235a-647b-404a-b20e-d39061ee4083