webif: Create webif directory with all the templates.
This commit prepares OSCam for automatic generation of module-webif-pages.c

All templates are extracted in webif/ directory, which holds pages_index.txt
and pages_gen.c which are used to generate module-webif-pages.c equivallent
from the extracted files.

The resulting webif/pages.c file when build results in exactly the same
object file as module-webif-pages.c, cleanups and extras would be added later.

Automatic building is not yet wired to OSCam build system, this would happen
in the next commit.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@8418 9dab235a-647b-404a-b20e-d39061ee4083