webif: Compress built-in templates.
Use the minilzo code that we already have to compress the built-in
templates. This saves ~90k binary size on 32bit defconfig. As the
tool reports:

> Compressed 148820 template bytes into 51619 bytes. 97201 saved bytes (65.31%).

The old code that works with non-compressed templates is still here,
to use it, comment `#define USE_COMPRESSION 1` in webif/pages_gen.c
and recompile.

Once we establish that template compression works fine the old
code can be removed.

This commit introduces no user visible changes.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@8446 9dab235a-647b-404a-b20e-d39061ee4083