log: Remove temporary log blocking.
Log blocking was introduced in r6585 as a way to control how much records
log_list can accumulate between flushes. The mechanism was needed because
logging thread slept 250 ms between processing the logging messages and in
that time a lot of message can be logged.

Now that we wake up logging thread on demand and do not stop processing
messages until the log_list is empty we can use much simpler mechanism
for making sure log_list is not getting too big.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@8595 9dab235a-647b-404a-b20e-d39061ee4083