Disable CLOCKFIX completely for now (doesn't matter if CLOCKFIX config is set).
We have reported regressions and too many moving parts for my taste.
This commit temporary disables CLOCKFIX features completely. It is
like before the first clockfix patch of the series was commited.

Please everybody test this (without applying any patches on top
of it) and see if your readers work the same way as in r9277.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@9308 9dab235a-647b-404a-b20e-d39061ee4083