dvbapi/samygo: Clean the code for future development.
According to the developers SAMYGO dvb code is not currently
working. This patch removes the config symbol and cleans the
code so future development can fix the code easily.

For more info: http://www.streamboard.tv/wbb2/thread.php?postid=533219

In theory if the code was working this transformation should
not break it and fixing only the samygo detection routine
should return it to the state before this patch minus all the
ifdefs.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@10214 9dab235a-647b-404a-b20e-d39061ee4083