Revert "Extra dump functions."
This reverts commit r10016.

Gratiuous copy+paste for no reason at all. These functions were not
even used. Not worth the extra 75 lines of code that does exactly
nothing (luckily the linker knows when a code is not used so these
functions never increased the binary size and after removing them
the binary is exactly the same).

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@10237 9dab235a-647b-404a-b20e-d39061ee4083