Allocate FTAB dynamically
The FTAB structure size is bigger than 2100 bytes and also the number
of filters is limited to 16.

This structure is used to implement the account filters (via 'ident' and
'chid' settings) and reader filters (via 'fallback_percaid', 'localcards',
'ident' and 'chid' settings).

These settings are not used by everyone but everyone takes the memory hit.

This patch makes the bulk of the structure dynammically allocated. The
memory savings are very nice:

{{{
 struct         before   after   delta  savings
  s_ftab          2116      16   -2100    99.3%
  s_reader       13592    5200   -8392    61.8%
  s_client        8120    3928   -4192    51.7%
  s_auth          6176    1976   -4200    68.1%

 Binary size:
  add/remove: 2/0 grow/shrink: 13/23 up/down: 1139/-434 (705)
}}}

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@10426 9dab235a-647b-404a-b20e-d39061ee4083