Add test module for oscam config entries parsers and generators.
We have lots of custom parsers and generators for our config file
settings. Unfortunately there are no test vectors for them or any
kind of test infrastructure which makes improving them or adding
features much harder that it needs to be.

This commit introduces stand alone `tests` module. The purpose of
this module (the code is in `tests.c`) is to collect test vectors
and to test them.

The module is stand-alone and it is built as separate binary by
using `make tests`.

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@10478 9dab235a-647b-404a-b20e-d39061ee4083