Replace cs_log() with rdr_log() where we have reader context.
This is not a complete solution, I only changed the places where
reader desc was already used in the logging message. There are
a lot of places (in modules) where we have reader context but
we're using cs_log(). If somebody wants to replace such cs_log()
calls with rdr_log(), grep for D_READER in modules-*.c

git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@10558 9dab235a-647b-404a-b20e-d39061ee4083