Fix setting client_proto_version for the connection
The client_proto_version was lost when the PMT data was not send in the very first data packet processed.

The trace for the problem:
- new socket connection fd: 14
- dvbapi_handlesockdata, client_proto_ver=0, pfd2[i]: i=0
- Got DVBAPI_CLIENT_INFO packet with size 42
- Client connected: 'vdr-plugin-dvbapi 2.2.1 / VDR 2.0.0' (protocol version = 2)
- dvbapi_handlesockdata AFTER, client_proto_ver=2, pfd2[i]: i=0
                                                
Now the connection is added to unassoc_fd, because there was only a greeting and no caPMT.
Then the PMT update is on other pfd2 index (i=1) and the client_proto was set previously on index i=0 !
- dvbapi_handlesockdata, client_proto_ver=0, pfd2[i]: i=1
                                                
And OSCam was using wrong protocol leading to problems with communication.

The commit uses the last_client_proto in this case to correctly fill the client_proto_version[] table with correct index.


git-svn-id: http://www.oscam.cc/svn/oscam-mirror/trunk@11209 9dab235a-647b-404a-b20e-d39061ee4083